+++ mayan/apps/templating/dependencies.py
@@ -1,11 +1,5 @@
-from mayan.apps.dependencies.classes import (
-    JavaScriptDependency, PythonDependency
-)
+from mayan.apps.dependencies.classes import JavaScriptDependency
 
 JavaScriptDependency(
     module=__name__, name='@highlightjs/cdn-assets', version_string='=11.9.0'
-)
-
-PythonDependency(
-    module=__name__, name='django-mathfilters', version_string='==1.0.0'
 )