+++ mayan/apps/templating/template_backends.py
@@ -14,7 +14,6 @@ class Template:
                     'BACKEND': 'django.template.backends.django.DjangoTemplates',
                     'OPTIONS': {
                         'builtins': [
-                            'mathfilters.templatetags.mathfilters',
                             'mayan.apps.templating.templatetags.templating_datetime_tags',
                             'mayan.apps.templating.templatetags.templating_json_tags',
                             'mayan.apps.templating.templatetags.templating_math_tags',