+++ HISTORY.rst
@@ -61,6 +61,7 @@
 - Remove setting `COMMON_COLLAPSE_LIST_MENU_LIST_FACET`. The behavior is
   now set as if the value of this setting was `True`.
 - Remove support for user themes.
+- Remove `django-mathfilters`.
 
 4.9.2 (20XX-XX-XX)
 ==================